pyproject.toml
[tool.poetry]
name = "quant-research-robot"
version = "0.1.0"
description = "基于LLM的能力自动复现量化领域的研究报告"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
litellm = "^1.0.0"
vnpy = "^4.0.0"
pandas = "^1.3.0"
numpy = "^1.20.0"
matplotlib = "^3.4.0"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
transformers = "^4.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.1.0"
isort = "^5.12.0"

[tool.poetry.scripts]
quant-robot = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 